Add async implementation of exec_run and file copy
