+++ services/orchest-ctl/app/docker_wrapper.py
@@ -1,0 +1,418 @@
+import asyncio
+import logging
+from typing import Any, Dict, Iterable, List, Literal, Optional, Tuple, Union
+
+import aiodocker
+import docker
+from docker.client import DockerClient
+from tqdm.asyncio import tqdm
+
+from app import utils
+from app.config import WRAP_LINES
+
+logger = logging.getLogger(__name__)
+
+
+class DockerWrapper:
+    def __init__(self):
+        self.sclient = DockerClient.from_env()  # sync
+        self._aclient: Optional[aiodocker.Docker] = None  # async
+
+    async def close_aclient(self):
+        if self._aclient is not None:
+            await self._aclient.close()
+            self._aclient = None
+
+    @property
+    def aclient(self):
+        if self._aclient is None:
+            self._aclient = aiodocker.Docker()
+
+        return self._aclient
+
+    def __del__(self):
+        # Called when the object is about to be destroyed.
+        self.sclient.close()
+
+    def is_network_installed(self, network: str) -> bool:
+        """Returns whether the given network is installed."""
+        try:
+            self.sclient.networks.get(network)
+        except docker.errors.NotFound:
+            return False
+        except docker.errors.APIError as e:
+            # TODO: I don't really know whether we want to catch this
+            #       error or just let it fail. We cannot let it fail
+            #       silently and so I see no need to log it as well
+            #       (orchest-ctl is auto-removed).
+
+            # Possible causes can be that the network already exists
+            # multiple times, making the request ambiguous.
+            logger.error("Unknow error when checking the network: {e}")
+
+            # TODO: raise custom error here
+            raise docker.errors.APIError(e)
+
+        return True
+
+    def install_network(self, network: str) -> None:
+        self.sclient.networks.create(network, driver="bridge")
+
+    async def _pull_image(self, image: str, force: bool = False) -> None:
+        """Pulls an image.
+
+        Args:
+            image: The image to pull.
+            force: If True, then pulls the image even when it is already
+                present and thus replacing it.
+
+        """
+        pull_image = False
+        try:
+            await self.aclient.images.get(image)
+        except aiodocker.exceptions.DockerError:
+            # Image is not yet installed.
+            pull_image = True
+
+        if force or pull_image:
+            await self.aclient.images.pull(image)
+
+    async def _pull_images(
+        self,
+        images: Iterable[str],
+        prog_bar: bool = True,
+        force: bool = False,
+    ):
+        pulls = [self._pull_image(image, force=force) for image in images]
+
+        if prog_bar:
+            pulls = tqdm.as_completed(
+                pulls,
+                total=len(pulls),
+                ncols=WRAP_LINES,
+                desc="Pulling images",
+                ascii=True,
+                position=0,
+                leave=True,
+                bar_format="{desc}: {n}/{total}|{bar}|",
+            )
+
+        for pull in pulls:
+            await pull
+
+        if prog_bar:
+            pulls.close()  # type: ignore
+
+            # Makes the next echo start on the line underneath the
+            # status bar instead of after.
+            await asyncio.sleep(0.05)
+            # TODO: Check whether we can use print(flush=True) here to
+            #       make this class not dependend on utils.
+            utils.echo()
+
+        await self.close_aclient()
+
+    def pull_images(
+        self,
+        images: Iterable[str],
+        prog_bar: bool = True,
+        force: bool = False,
+    ):
+        """Pulls an iterable of images.
+
+        Args:
+            images: The images to pull.
+            prog_bar: Whether or not to show a progress bar to
+                indicate progress.
+            force: If True, then pulls an image even when it is already
+                present and thus replacing it.
+
+        """
+        return asyncio.run(self._pull_images(images, prog_bar=prog_bar, force=force))
+
+    async def _does_image_exist(self, image: str) -> bool:
+        try:
+            await self.aclient.images.inspect(image)
+        except aiodocker.exceptions.DockerError:
+            return False
+
+        return True
+
+    async def _do_images_exist(self, images: Iterable[str]) -> List[bool]:
+        res = await asyncio.gather(*[self._does_image_exist(image) for image in images])
+        await self.close_aclient()
+
+        return res
+
+    def do_images_exist(self, images: Iterable[str]) -> List[bool]:
+        """Checks whether the given images exist.
+
+        True if found, False is not. Order in returned list equals
+        order of given `images`.
+        """
+        return asyncio.run(self._do_images_exist(images))
+
+    async def _list_image_ids(self, all: bool = False, label: Optional[str] = None):
+        images = await self.aclient.images.list(all=all, filters={"label": [label]})
+        await self.close_aclient()
+
+        return [img["Id"] for img in images]
+
+    def list_image_ids(self, all: bool = False, label: Optional[str] = None):
+        return asyncio.run(self._list_image_ids(all=all, label=label))
+
+    async def _remove_image(self, image: Iterable[str], force: bool = False):
+        await self.aclient.images.delete(image, force=force)
+
+    async def _remove_images(self, image_ids: Iterable[str], force: bool = False):
+        await asyncio.gather(
+            *[self._remove_image(img, force=force) for img in image_ids]
+        )
+        await self.close_aclient()
+
+    def remove_images(self, image_ids: Iterable[str], force: bool = False):
+        """
+
+        Args:
+            image_ids: Iterable of image IDs.
+            force: Remove an image even if it is being used by stopped
+                   containers or has other tags.
+
+        """
+        # TODO: use typing to state that str should be of type ID
+        asyncio.run(self._remove_images(image_ids, force=force))
+
+    async def _get_containers(
+        self,
+        state: Literal["all", "running", "exited"] = "running",
+        network: Optional[str] = None,
+    ) -> Tuple[List[str], List[Optional[str]]]:
+        all_ = True if state in ["all", "exited"] else False
+        containers = await self.aclient.containers.list(
+            all=all_, filters={"network": [network]}
+        )
+        await self.close_aclient()
+
+        ids = []
+        img_names = []
+        for c in containers:
+            if state == "exited":
+                if c._container.get("State") != "running":
+                    img_names.append(c._container.get("Image"))
+                    ids.append(c.id)
+
+            else:
+                img_names.append(c._container.get("Image"))
+                ids.append(c.id)
+
+        return ids, img_names
+
+    def get_containers(
+        self,
+        state: Literal["all", "running", "exited"] = "running",
+        network: Optional[str] = None,
+    ) -> Tuple[List[str], List[Optional[str]]]:
+        """Returns runnings containers (on a network).
+
+        Args:
+            state: The state of the container to be in in order for it
+                to be returned.
+            network: The network on which to filter the containers.
+
+        Returns:
+            (container_ids, img_names) in respective order. Where the
+            img_names are the names of the images underlying of the
+            containers.
+
+        """
+        return asyncio.run(self._get_containers(state=state, network=network))
+
+    async def _remove_containers(self, container_ids: Iterable[str]):
+        async def remove_container(id_: str) -> None:
+            container = self.aclient.containers.container(id_)
+
+            # If the container is running, kill it before removing it
+            # and remove anonymous volumes associated with the
+            # container.
+            await container.delete(force=True, v=True)
+
+            try:
+                # Block until the container is deleted.
+                await container.wait(condition="removed")
+            except aiodocker.exceptions.DockerError:
+                # The container was removed so fast that the wait
+                # condition was unable to find the container.
+                pass
+
+        await asyncio.gather(*[remove_container(id_) for id_ in container_ids])
+        await self.close_aclient()
+
+    def remove_containers(self, container_ids: Iterable[str]):
+        """Removes the given containers and removes associated volumes.
+
+        If the container is running, then it is killed before removing
+        it.
+
+        Args:
+            containers: An iterable of containers, where a container
+                is given by its name or ID.
+
+        """
+        asyncio.run(self._remove_containers(container_ids))
+
+    async def _run_container(self, name, config, use_name=False, detach=True):
+        if use_name:
+            container = await self.aclient.containers.run(name=name, config=config)
+        else:
+            container = await self.aclient.containers.run(config=config)
+
+        stdout = None
+        if not detach:
+            await container.wait()
+            stdout = await container.log(stdout=True)
+
+        info = {
+            "id": container.id,
+            "stdout": stdout,
+        }
+        return name, info
+
+    async def _run_containers(
+        self, configs: Dict[str, Dict[str, Any]], use_name=False, detach=True
+    ):
+        stdouts = await asyncio.gather(
+            *[
+                self._run_container(name, config, use_name=use_name, detach=detach)
+                for name, config in configs.items()
+            ]
+        )
+
+        await self.close_aclient()
+
+        return dict(stdouts)
+
+    def run_containers(
+        self, configs: Dict[str, Dict[str, Any]], use_name=False, detach=True
+    ):
+        """Runs the given collection of images.
+
+        Args:
+            configs: Configuration to start each container with.
+                Example:
+                    {
+                      "orchest-api": {...}
+                    }
+            use_name: If True uses the keys of `configs` as the names
+                for the Docker containers. Otherwise it lets Docker
+                choose a name. NOTE: Name collissions will prevent
+                containers from starting.
+
+        """
+        return asyncio.run(
+            self._run_containers(configs, use_name=use_name, detach=detach)
+        )
+
+    def get_container(self, id_or_name: str):
+        return self.sclient.containers.get(id_or_name)
+
+    def exec_run(self, container_id: str, cmd: Union[str, List[Any]]) -> int:
+        """Returns the exit code of running a cmd inside a container.
+
+        Making this function async (rough idea):
+        ```python
+        container = self.aclient.containers.container(container_id)
+
+        # If the container is running, kill it before removing it
+        # and remove anonymous volumes associated with the
+        # container.
+        exec = await container.exec(...)
+
+        https://github.com/aio-libs/aiodocker/blob/master/aiodocker/execs.py#L69
+        exec.start(...)
+
+        # Not sure whether start has to be called before inspect
+        https://github.com/aio-libs/aiodocker/blob/master/aiodocker/execs.py#L34
+        res = exec.inspect(...)
+
+        https://docs.docker.com/engine/api/v1.41/#operation/ExecInspect
+        res["ExitCode"]
+        ```
+        """
+        container = docker.models.containers.Container(
+            attrs={"Id": container_id}, client=self.sclient
+        )
+        # Calling the lower level API as the following command gave
+        # incorrect exit codes:
+        # exit_code, _ = container.exec_run(cmd)
+        resp = container.client.api.exec_create(container.id, cmd)
+        _ = container.client.api.exec_start(resp["Id"])
+        exit_code = container.client.api.exec_inspect(resp["Id"])["ExitCode"]
+
+        return exit_code
+
+    async def _exec_run(self, container_id: str, cmd) -> int:
+        """Returns the exit code of running a cmd inside a container."""
+
+        container = self.aclient.containers.container(container_id)
+
+        exec = await container.exec(stdout=True, stderr=True, stdin=True, cmd=cmd)
+
+        await exec.start(detach=True)
+
+        exit_code = None
+        while exit_code is None:
+            exit_code = (await exec.inspect())["ExitCode"]
+            await asyncio.sleep(0.1)
+
+        return exit_code
+
+    async def _exec_runs(self, cmds):
+        res = await asyncio.gather(*[self._exec_run(id, cmd) for id, cmd in cmds])
+        await self.close_aclient()
+        return res
+
+    def exec_runs(self, cmds: List[Tuple[str, Union[str, List[str]]]]) -> List[int]:
+        """Returns the exit codes of running cmds in containers.
+
+        Args:
+            cmds: A list of pairs where, for each pair, the first
+            element is the container id, while the second element is the
+            command to run.
+
+        Returns:
+            List of exit codes of the commands that have been run.
+        """
+        return asyncio.run(self._exec_runs(cmds))
+
+    async def _copy_file_from_container(
+        self, container_id: str, from_path: str, to_path: str
+    ):
+        container = self.aclient.containers.container(container_id)
+
+        tar_file = await container.get_archive(from_path)
+        file_name = from_path if "/" not in from_path else from_path.split("/")[-1]
+        file = tar_file.extractfile(file_name)
+
+        with open(to_path, "wb") as dest_file:
+            dest_file.write(file.read())
+        tar_file.close()
+
+    async def _copy_file_from_containers(self, files: List[Tuple[str, str, str]]):
+        await asyncio.gather(
+            *[
+                self._copy_file_from_container(id, from_path, to_path)
+                for (id, from_path, to_path) in files
+            ]
+        )
+        await self.close_aclient()
+
+    def copy_files_from_containers(self, files: List[Tuple[str, str, str]]):
+        """AI is creating summary for copy_files_from_containers
+
+        Args:
+            files: A list of triples where, for each triple, the first
+            element is the container id, the second element is a path
+            in the container (to copy from) and the third element is a
+            path in the fs (to copy to).
+        """
+        asyncio.run(self._copy_file_from_containers(files))
