+++ services/orchest-ctl/app/orchest.py
@@ -9,513 +9,22 @@ TODO:
       functionality. See: https://www.python.org/dev/peps/pep-0585/
 
 """
-import asyncio
-import io
-import json
 import logging
 import os
-import tarfile
 from functools import reduce
-from typing import Any, Dict, Iterable, List, Literal, Optional, Set, Tuple, Union
-
-import aiodocker
-import docker
-from docker.client import DockerClient
-from tqdm.asyncio import tqdm
+from typing import List, Optional, Set, Tuple
 
 from app import spec, utils
-from app.config import DOCKER_NETWORK, ORCHEST_IMAGES, WRAP_LINES
+from app.config import ORCHEST_IMAGES, WRAP_LINES, _on_start_images
+from app.debug import debug_dump, health_check
+from app.docker_wrapper import DockerWrapper
+from app.orchest_resource_manager import OrchestResourceManager
 
 logger = logging.getLogger(__name__)
 
 
-class DockerWrapper:
-    def __init__(self):
-        self.sclient = DockerClient.from_env()  # sync
-        self._aclient: Optional[aiodocker.Docker] = None  # async
-
-    async def close_aclient(self):
-        if self._aclient is not None:
-            await self._aclient.close()
-            self._aclient = None
-
-    @property
-    def aclient(self):
-        if self._aclient is None:
-            self._aclient = aiodocker.Docker()
-
-        return self._aclient
-
-    def __del__(self):
-        # Called when the object is about to be destroyed.
-        self.sclient.close()
-
-    def is_network_installed(self, network: str) -> bool:
-        """Returns whether the given network is installed."""
-        try:
-            self.sclient.networks.get(network)
-        except docker.errors.NotFound:
-            return False
-        except docker.errors.APIError as e:
-            # TODO: I don't really know whether we want to catch this
-            #       error or just let it fail. We cannot let it fail
-            #       silently and so I see no need to log it as well
-            #       (orchest-ctl is auto-removed).
-
-            # Possible causes can be that the network already exists
-            # multiple times, making the request ambiguous.
-            logger.error("Unknow error when checking the network: {e}")
-
-            # TODO: raise custom error here
-            raise docker.errors.APIError(e)
-
-        return True
-
-    def install_network(self, network: str) -> None:
-        self.sclient.networks.create(network, driver="bridge")
-
-    async def _pull_image(self, image: str, force: bool = False) -> None:
-        """Pulls an image.
-
-        Args:
-            image: The image to pull.
-            force: If True, then pulls the image even when it is already
-                present and thus replacing it.
-
-        """
-        pull_image = False
-        try:
-            await self.aclient.images.get(image)
-        except aiodocker.exceptions.DockerError:
-            # Image is not yet installed.
-            pull_image = True
-
-        if force or pull_image:
-            await self.aclient.images.pull(image)
-
-    async def _pull_images(
-        self,
-        images: Iterable[str],
-        prog_bar: bool = True,
-        force: bool = False,
-    ):
-        pulls = [self._pull_image(image, force=force) for image in images]
-
-        if prog_bar:
-            pulls = tqdm.as_completed(
-                pulls,
-                total=len(pulls),
-                ncols=WRAP_LINES,
-                desc="Pulling images",
-                ascii=True,
-                position=0,
-                leave=True,
-                bar_format="{desc}: {n}/{total}|{bar}|",
-            )
-
-        for pull in pulls:
-            await pull
-
-        if prog_bar:
-            pulls.close()  # type: ignore
-
-            # Makes the next echo start on the line underneath the
-            # status bar instead of after.
-            await asyncio.sleep(0.05)
-            # TODO: Check whether we can use print(flush=True) here to
-            #       make this class not dependend on utils.
-            utils.echo()
-
-        await self.close_aclient()
-
-    def pull_images(
-        self,
-        images: Iterable[str],
-        prog_bar: bool = True,
-        force: bool = False,
-    ):
-        """Pulls an iterable of images.
-
-        Args:
-            images: The images to pull.
-            prog_bar: Whether or not to show a progress bar to
-                indicate progress.
-            force: If True, then pulls an image even when it is already
-                present and thus replacing it.
-
-        """
-        return asyncio.run(self._pull_images(images, prog_bar=prog_bar, force=force))
-
-    async def _does_image_exist(self, image: str) -> bool:
-        try:
-            await self.aclient.images.inspect(image)
-        except aiodocker.exceptions.DockerError:
-            return False
-
-        return True
-
-    async def _do_images_exist(self, images: Iterable[str]) -> List[bool]:
-        res = await asyncio.gather(*[self._does_image_exist(image) for image in images])
-        await self.close_aclient()
-
-        return res
-
-    def do_images_exist(self, images: Iterable[str]) -> List[bool]:
-        """Checks whether the given images exist.
-
-        True if found, False is not. Order in returned list equals
-        order of given `images`.
-        """
-        return asyncio.run(self._do_images_exist(images))
-
-    async def _list_image_ids(self, all: bool = False, label: Optional[str] = None):
-        images = await self.aclient.images.list(all=all, filters={"label": [label]})
-        await self.close_aclient()
-
-        return [img["Id"] for img in images]
-
-    def list_image_ids(self, all: bool = False, label: Optional[str] = None):
-        return asyncio.run(self._list_image_ids(all=all, label=label))
-
-    async def _remove_image(self, image: Iterable[str], force: bool = False):
-        await self.aclient.images.delete(image, force=force)
-
-    async def _remove_images(self, image_ids: Iterable[str], force: bool = False):
-        await asyncio.gather(
-            *[self._remove_image(img, force=force) for img in image_ids]
-        )
-        await self.close_aclient()
-
-    def remove_images(self, image_ids: Iterable[str], force: bool = False):
-        """
-
-        Args:
-            image_ids: Iterable of image IDs.
-            force: Remove an image even if it is being used by stopped
-                   containers or has other tags.
-
-        """
-        # TODO: use typing to state that str should be of type ID
-        asyncio.run(self._remove_images(image_ids, force=force))
-
-    async def _get_containers(
-        self,
-        state: Literal["all", "running", "exited"] = "running",
-        network: Optional[str] = None,
-    ) -> Tuple[List[str], List[Optional[str]]]:
-        all_ = True if state in ["all", "exited"] else False
-        containers = await self.aclient.containers.list(
-            all=all_, filters={"network": [network]}
-        )
-        await self.close_aclient()
-
-        ids = []
-        img_names = []
-        for c in containers:
-            if state == "exited":
-                if c._container.get("State") != "running":
-                    img_names.append(c._container.get("Image"))
-                    ids.append(c.id)
-
-            else:
-                img_names.append(c._container.get("Image"))
-                ids.append(c.id)
-
-        return ids, img_names
-
-    def get_containers(
-        self,
-        state: Literal["all", "running", "exited"] = "running",
-        network: Optional[str] = None,
-    ) -> Tuple[List[str], List[Optional[str]]]:
-        """Returns runnings containers (on a network).
-
-        Args:
-            state: The state of the container to be in in order for it
-                to be returned.
-            network: The network on which to filter the containers.
-
-        Returns:
-            (container_ids, img_names) in respective order. Where the
-            img_names are the names of the images underlying of the
-            containers.
-
-        """
-        return asyncio.run(self._get_containers(state=state, network=network))
-
-    async def _remove_containers(self, container_ids: Iterable[str]):
-        async def remove_container(id_: str) -> None:
-            container = self.aclient.containers.container(id_)
-
-            # If the container is running, kill it before removing it
-            # and remove anonymous volumes associated with the
-            # container.
-            await container.delete(force=True, v=True)
-
-            try:
-                # Block until the container is deleted.
-                await container.wait(condition="removed")
-            except aiodocker.exceptions.DockerError:
-                # The container was removed so fast that the wait
-                # condition was unable to find the container.
-                pass
-
-        await asyncio.gather(*[remove_container(id_) for id_ in container_ids])
-        await self.close_aclient()
-
-    def remove_containers(self, container_ids: Iterable[str]):
-        """Removes the given containers and removes associated volumes.
-
-        If the container is running, then it is killed before removing
-        it.
-
-        Args:
-            containers: An iterable of containers, where a container
-                is given by its name or ID.
-
-        """
-        asyncio.run(self._remove_containers(container_ids))
-
-    async def _run_container(self, name, config, use_name=False, detach=True):
-        if use_name:
-            container = await self.aclient.containers.run(name=name, config=config)
-        else:
-            container = await self.aclient.containers.run(config=config)
-
-        stdout = None
-        if not detach:
-            await container.wait()
-            stdout = await container.log(stdout=True)
-
-        info = {
-            "id": container.id,
-            "stdout": stdout,
-        }
-        return name, info
-
-    async def _run_containers(
-        self, configs: Dict[str, Dict[str, Any]], use_name=False, detach=True
-    ):
-        stdouts = await asyncio.gather(
-            *[
-                self._run_container(name, config, use_name=use_name, detach=detach)
-                for name, config in configs.items()
-            ]
-        )
-
-        await self.close_aclient()
-
-        return dict(stdouts)
-
-    def run_containers(
-        self, configs: Dict[str, Dict[str, Any]], use_name=False, detach=True
-    ):
-        """Runs the given collection of images.
-
-        Args:
-            configs: Configuration to start each container with.
-                Example:
-                    {
-                      "orchest-api": {...}
-                    }
-            use_name: If True uses the keys of `configs` as the names
-                for the Docker containers. Otherwise it lets Docker
-                choose a name. NOTE: Name collissions will prevent
-                containers from starting.
-
-        """
-        return asyncio.run(
-            self._run_containers(configs, use_name=use_name, detach=detach)
-        )
-
-    def exec_run(self, container_id: str, cmd: Union[str, List[Any]]) -> int:
-        """Returns the exit code of running a cmd inside a container.
-
-        Making this function async (rough idea):
-        ```python
-        container = self.aclient.containers.container(container_id)
-
-        # If the container is running, kill it before removing it
-        # and remove anonymous volumes associated with the
-        # container.
-        exec = await container.exec(...)
-
-        https://github.com/aio-libs/aiodocker/blob/master/aiodocker/execs.py#L69
-        exec.start(...)
-
-        # Not sure whether start has to be called before inspect
-        https://github.com/aio-libs/aiodocker/blob/master/aiodocker/execs.py#L34
-        res = exec.inspect(...)
-
-        https://docs.docker.com/engine/api/v1.41/#operation/ExecInspect
-        res["ExitCode"]
-        ```
-        """
-        container = docker.models.containers.Container(
-            attrs={"Id": container_id}, client=self.sclient
-        )
-        # Calling the lower level API as the following command gave
-        # incorrect exit codes:
-        # exit_code, _ = container.exec_run(cmd)
-        resp = container.client.api.exec_create(container.id, cmd)
-        _ = container.client.api.exec_start(resp["Id"])
-        exit_code = container.client.api.exec_inspect(resp["Id"])["ExitCode"]
-
-        return exit_code
-
-    def get_container(self, id_or_name: str):
-        return self.sclient.containers.get(id_or_name)
-
-    def copy_container_file(self, container, from_path: str, to_path: str):
-        """Copy a file from an existing container to a destination.
-
-        Args:
-            container:
-            from:
-            to:
-
-        """
-        # Get bytes from container.
-        data, stat = container.get_archive(from_path)
-        filelike = io.BytesIO(b"".join(b for b in data))
-
-        # Untar so that we can write it as a "normal" file
-        tar = tarfile.open(fileobj=filelike)
-        file = tar.extractfile(stat["name"])
-
-        with open(to_path, "wb") as dest_file:
-            dest_file.write(file.read())
-        filelike.close()
-        tar.close()
-
-
-class OrchestResourceManager:
-    orchest_images: List[str] = ORCHEST_IMAGES["all"]
-    network: str = DOCKER_NETWORK
-
-    def __init__(self):
-        self.docker_client = DockerWrapper()
-
-    def install_network(self) -> None:
-        """Installs the Orchest Docker network."""
-        # Don't install the network again if it is already installed
-        # because that will create the another network with the same
-        # name but with another ID. Thereby, breaking Orchest.
-        try:
-            is_installed = self.docker_client.is_network_installed(self.network)
-        except docker.errors.APIError:
-            # TODO: reraise the error but with a helpful message that
-            # helps the user fix the issue.
-            raise
-
-        if not is_installed:
-            # We only want to print this message to the user once. The
-            # best bet is that if the Orchest network has not yet been
-            # installed, then most likely the user has not seen this
-            # message before.
-            utils.echo(
-                "Orchest sends anonymized telemetry to analytics.orchest.io."
-                " To disable it, please refer to:",
-                wrap=WRAP_LINES,
-            )
-            utils.echo(
-                "\thttps://orchest.readthedocs.io/en/stable/user_guide/other.html#configuration"  # noqa: E501, W505
-            )
-
-            self.docker_client.install_network(self.network)
-
-    def get_images(self, orchest_owned: bool = False) -> List[str]:
-        """Returns all pulled images associated to Orchest.
-
-        Args:
-            orchest_owned: If True only returns the images owned by the
-                Orchest organization, e.g. excluding "rabbitmq".
-
-        """
-        check_images = self.orchest_images
-        if orchest_owned:
-            check_images = [
-                img for img in self.orchest_images if img.startswith("orchest")
-            ]
-
-        exists = self.docker_client.do_images_exist(check_images)
-
-        # TODO: could make it into set as well as order is not important
-        #       here.
-        return [img for i, img in enumerate(check_images) if exists[i]]
-
-    # TODO: this function might be a bit strange if it
-    #       returns img names.
-    def get_containers(
-        self,
-        state: Literal["all", "running", "exited"] = "running",
-    ) -> Tuple[List[str], List[Optional[str]]]:
-        """
-
-        Args:
-            state: The state of the container to be in in order for it
-                to be returned.
-        """
-        return self.docker_client.get_containers(state=state, network=self.network)
-
-    def get_env_build_imgs(self):
-        return self.docker_client.list_image_ids(label="_orchest_project_uuid")
-
-    def remove_env_build_imgs(self):
-        env_build_imgs = self.get_env_build_imgs()
-        self.docker_client.remove_images(env_build_imgs, force=True)
-
-
 class OrchestApp:
-    """...
-
-    Attributes:
-        on_start_images: Images to run when the app is started. The
-            order states the order in which the images have to be
-            started due to dependencies between them. A collection
-            indicates that its contained images can be started
-            asynchronously.
-
-    """
-
-    # postgres -> orchest-webserver, orchest-api, auth-server
-    # rabbitmq -> celery-worker
-    # ... -> nginx-proxy (otherwise user gets error 500)
-    on_start_images: List[Set[str]] = [
-        set(
-            [
-                "postgres:13.1",
-                "orchest/file-manager:latest",
-                "rabbitmq:3",
-            ]
-        ),
-        set(
-            [
-                "orchest/orchest-api:latest",
-                "orchest/orchest-webserver:latest",
-                "orchest/celery-worker:latest",
-                "orchest/auth-server:latest",
-            ]
-        ),
-        set(
-            [
-                "orchest/nginx-proxy:latest",
-            ]
-        ),
-    ]
-
-    health_check_command = {
-        # Works locally, doesn't work here.
-        # "orchest/nginx-proxy:latest": 'service --status-all |
-        # grep openresty',
-        "orchest/orchest-api:latest": "wget localhost/api --spider",
-        "orchest/orchest-webserver:latest": "wget localhost --spider",
-        "orchest/auth-server:latest": "wget localhost/auth --spider",
-        "orchest/celery-worker:latest": "celery inspect ping -A app.core.tasks",
-        "orchest/file-manager:latest": "wget localhost --spider",
-        "postgres:13.1": "pg_isready --username postgres",
-        "rabbitmq:3": "rabbitmqctl node_health_check",
-    }
+    """..."""
 
     def __init__(self):
         self.resource_manager = OrchestResourceManager()
@@ -524,8 +33,8 @@ class OrchestApp:
     def is_running(self, running_containers) -> bool:
         """Check whether Orchest is running"""
 
-        # Don't count orchest-ctl when checking
-        # whether Orchest is running.
+        # Don't count orchest-ctl when checking whether Orchest is
+        # running.
         running_containers = [
             c for c in running_containers if c not in ["orchest/orchest-ctl:latest"]
         ]
@@ -578,9 +87,7 @@ class OrchestApp:
         # Check whether the minimal set of images is present for Orchest
         # to be started.
         pulled_images = self.resource_manager.get_images()
-        req_images: Set[str] = reduce(
-            lambda x, y: x.union(y), self.on_start_images, set()
-        )
+        req_images: Set[str] = reduce(lambda x, y: x.union(y), _on_start_images, set())
         missing_images = req_images - set(pulled_images)
 
         if missing_images or not self.docker_client.is_network_installed(
@@ -632,7 +139,7 @@ class OrchestApp:
 
         # Start the containers in the correct order, keeping in mind
         # dependencies between containers.
-        for i, to_start_imgs in enumerate(self.on_start_images):
+        for i, to_start_imgs in enumerate(_on_start_images):
             filter_ = {"Image": to_start_imgs}
             config = spec.filter_container_config(container_config, filter=filter_)
             stdouts = self.docker_client.run_containers(
@@ -715,7 +222,7 @@ class OrchestApp:
 
         self.docker_client.run_containers(config, use_name=True, detach=True)
 
-    def status(self):
+    def status(self, ext=False):
 
         _, running_containers_names = self.resource_manager.get_containers(
             state="running"
@@ -727,9 +234,7 @@ class OrchestApp:
 
         # Minimal set of containers to be running for Orchest to be in
         # a valid state.
-        valid_set: Set[str] = reduce(
-            lambda x, y: x.union(y), self.on_start_images, set()
-        )
+        valid_set: Set[str] = reduce(lambda x, y: x.union(y), _on_start_images, set())
 
         if valid_set - set(running_containers_names):
             utils.echo("Orchest is running, but has reached an invalid state. Run:")
@@ -740,244 +245,16 @@ class OrchestApp:
             )
         else:
             utils.echo("Orchest is running.")
-            for container, exit_code in self.healt_check().items():
-                if exit_code != 0:
-                    utils.echo(f"{container} is not ready ({exit_code}).")
+            if ext:
+                utils.echo("Performing extensive status checks...")
+                no_issues = True
+                for container, exit_code in health_check().items():
+                    if exit_code != 0:
+                        no_issues = False
+                        utils.echo(f"{container} is not ready ({exit_code}).")
 
-    def health_check(self):
-        """
-
-        Returns:
-            Dict mapping container names to health check exit codes.
-        """
-        (
-            running_containers_ids,
-            running_containers_names,
-        ) = self.resource_manager.get_containers(state="running")
-
-        exit_codes = {}
-        # Run health checks for containers that do have one.
-        for id, container in zip(running_containers_ids, running_containers_names):
-            hcheck = self.health_check_command.get(container)
-            if hcheck is None:
-                continue
-            exit_codes[container] = self.docker_client.exec_run(id, hcheck)
-        return exit_codes
-
-    def database_debug_dump(
-        self, path: str, dbs=["auth_server", "orchest_api", "orchest_webserver"]
-    ):
-        """Get database schema, revision version, rows per table(s)."""
-
-        container = self.docker_client.get_container("orchest-database")
-
-        db_debug_dump_directory = os.path.join(path, "database")
-        if not os.path.exists(db_debug_dump_directory):
-            os.mkdir(db_debug_dump_directory)
-
-        for db in dbs:
-
-            # Schema of the db.
-            filename = f"{db}_schema.sql"
-            container.exec_run(
-                (f"pg_dump --user postgres -d {db}" f" --schema-only -f {filename}")
-            )
-            self.docker_client.copy_container_file(
-                container, filename, os.path.join(db_debug_dump_directory, filename)
-            )
-
-            # Alembic version.
-            filename = f"{db}_alembic_version.sql"
-            cmd = (
-                f'psql --user postgres -d {db} -c "SELECT * '
-                f'FROM alembic_version;" -o {filename}'
-            )
-            container.exec_run(cmd)
-            self.docker_client.copy_container_file(
-                container, filename, os.path.join(db_debug_dump_directory, filename)
-            )
-
-            # Row counts of each table in the db.
-            filename = f"{db}_counts.sql"
-            cmd = (
-                f'psql --user postgres -d {db} -c "'
-                """
-                select table_schema, table_name,
-                    (xpath('/row/cnt/text()', xml_count))[1]::text::int as row_count
-                from (
-                    select table_name, table_schema,
-                    query_to_xml(format('select count(*) as cnt from %I.%I',
-                    table_schema, table_name), false, true, '') as xml_count
-                from information_schema.tables
-                where table_schema = 'public') t
-                """
-                f'" -o {filename}'
-            )
-            container.exec_run(cmd)
-            self.docker_client.copy_container_file(
-                container, filename, os.path.join(db_debug_dump_directory, filename)
-            )
-
-    def celery_debug_dump(self, path: str):
-        """Worker logs and output of celery inspect commands."""
-
-        container = self.docker_client.get_container("celery-worker")
-        cel_debug_dump_directory = os.path.join(path, "celery")
-
-        if not os.path.exists(cel_debug_dump_directory):
-            os.mkdir(cel_debug_dump_directory)
-
-        for worker in [
-            "celery_env_builds",
-            "celery_interactive",
-            "celery_jobs",
-        ]:
-            self.docker_client.copy_container_file(
-                container,
-                f"/orchest/services/orchest-api/app/{worker}.log",
-                os.path.join(cel_debug_dump_directory, f"{worker}.log"),
-            )
-
-        cmd_template = "celery inspect -A app.core.tasks {name} > {name}.txt"
-        for inspect_command in [
-            "active",
-            "active_queues",
-            "conf",
-            "revoked",
-            "report",
-            "reserved",
-            "scheduled",
-            "stats",
-        ]:
-            cmd = cmd_template.format(name=inspect_command)
-            # Necessary work around to avoid errors of the python docker
-            # SDK.
-            cmd = ["/bin/sh", "-c", cmd]
-            container.exec_run(cmd)
-            self.docker_client.copy_container_file(
-                container,
-                f"/orchest/services/orchest-api/app/{inspect_command}.txt",
-                os.path.join(cel_debug_dump_directory, f"{inspect_command}.txt"),
-            )
-
-    def websever_debug_dump(self, path):
-        """Get the webserver log file."""
-
-        container = self.docker_client.get_container("orchest-webserver")
-
-        web_debug_dump_directory = os.path.join(path, "webserver")
-        if not os.path.exists(web_debug_dump_directory):
-            os.mkdir(web_debug_dump_directory)
-
-        self.docker_client.copy_container_file(
-            container,
-            "/orchest/services/orchest-webserver/app/orchest-webserver.log",
-            os.path.join(web_debug_dump_directory, "orchest-webserver.log"),
-        )
-
-    def containers_logs_dump(self, path):
-        """Get the logs of every Orchest container, except steps."""
-
-        containers_logs = os.path.join(path, "containers-logs")
-        if not os.path.exists(containers_logs):
-            os.mkdir(containers_logs)
-
-        ids, names = self.resource_manager.get_containers(state="all")
-
-        orchest_set: Set[str] = reduce(
-            lambda x, y: x.union(y), self.on_start_images, set()
-        )
-        session_containers = {
-            "orchest/jupyter-server:latest",
-            "orchest/jupyter-enterprise-gateway",
-            "orchest/memory-server:latest",
-        }
-
-        for id, name in zip(ids, names):
-            if name in orchest_set:
-                file_name = f"{name}.txt"
-            elif name in session_containers:
-                file_name = f"{name}-{id}.txt"
-            # Do not pickup containers that are running user pipeline
-            # steps, to avoid the risk of getting user data through its
-            # logs.
-            else:
-                continue
-
-            # Else orchest/<something> won't work as a file name in
-            # os.path.join.
-            file_name = file_name.replace("orchest/", "")
-            container = self.docker_client.get_container(id)
-            logs = container.logs()
-
-            with open(os.path.join(containers_logs, file_name), "wb") as file:
-                file.write(logs)
-
-    def containers_version_dump(self, path):
-        """Get the version of Orchest containers"""
-
-        with open(os.path.join(path, "containers_version.txt"), "w") as file:
-            for name, version in self.containers_version().items():
-                file.write(f"{name:<44}: {version}\n")
-
-    def orchest_config_dump(self, path):
-        """Get the Orchest config file, with telemetry UUID removed"""
-
-        # Copy the config
-        with open("/config/config.json") as input_json_file:
-            config = json.load(input_json_file)
-            # Removed for privacy.
-            del config["TELEMETRY_UUID"]
-
-            with open(os.path.join(path, "config.json"), "w") as output_json_file:
-                json.dump(config, output_json_file)
-
-    def health_check_dump(self, path):
-        with open(os.path.join(path, "health_check.txt"), "w") as file:
-            for container, exit_code in self.health_check().items():
-                file.write(f"{container:<44}: {exit_code}\n")
-
-    def running_containers_dump(self, path):
-        """Get which Orchest containers are running"""
-
-        _, running_containers_names = self.resource_manager.get_containers(
-            state="running"
-        )
-        with open(os.path.join(path, "running_containers.txt"), "w") as file:
-            for name in running_containers_names:
-                file.write(f"{name}\n")
-
-    def debug_dump(self):
-
-        debug_dump_directory = "/orchest-host/debug-dump"
-        if not os.path.exists(debug_dump_directory):
-            os.mkdir(debug_dump_directory)
-        elif os.path.isfile(debug_dump_directory):
-            utils.echo("Expected debug-dump to be a directory, whereas it is a file.")
-            return
-
-        errors = []
-        for name, func in [
-            ("config", self.orchest_config_dump),
-            ("containers version", self.containers_version_dump),
-            ("containers logs", self.containers_logs_dump),
-            ("running containers", self.running_containers_dump),
-            ("health check", self.health_check_dump),
-            ("database", self.database_debug_dump),
-            ("celery", self.celery_debug_dump),
-            ("webserver", self.websever_debug_dump),
-        ]:
-            try:
-                func(debug_dump_directory)
-            except Exception as e:
-                errors.append((name, e))
-
-        with open(os.path.join(debug_dump_directory, "errors.txt"), "w") as file:
-            file.write(
-                "This is a log of errors that happened during the dump, if any.\n"
-            )
-            for name, exc in errors:
-                file.write(f"{name}: {exc}\n")
+                if no_issues:
+                    utils.echo("All services are ready.")
 
     def update(self, mode=None):
         """Update Orchest.
@@ -1039,20 +316,6 @@ class OrchestApp:
         utils.echo("Don't forget to restart Orchest for the changes to take effect:")
         utils.echo("\torchest restart")
 
-    def containers_version(self):
-        pulled_images = self.resource_manager.get_images(orchest_owned=True)
-        configs = {}
-        for img in pulled_images:
-            configs[img] = {
-                "Image": img,
-                "Entrypoint": ["printenv", "ORCHEST_VERSION"],
-            }
-
-        stdouts = self.docker_client.run_containers(configs, detach=False)
-        for img in stdouts.keys():
-            stdouts[img] = stdouts[img]["stdout"][0].rstrip()
-        return stdouts
-
     def version(self, ext=False):
         """Returns the version of Orchest.
 
@@ -1068,16 +331,8 @@ class OrchestApp:
             return
 
         utils.echo("Getting versions of all containers...")
-        pulled_images = self.resource_manager.get_images(orchest_owned=True)
 
-        configs = {}
-        for img in pulled_images:
-            configs[img] = {
-                "Image": img,
-                "Entrypoint": ["printenv", "ORCHEST_VERSION"],
-            }
-
-        stdouts = self.containers_version()
+        stdouts = self.resource_manager.containers_version()
         stdout_values = set()
         for img, stdout in stdouts.items():
             stdout_values.add(stdout)
@@ -1092,6 +347,9 @@ class OrchestApp:
             )
             utils.echo("\torchest update")
             utils.echo("To get all containers on the same version again.")
+
+    def debug(self, compress: bool):
+        debug_dump(compress)
 
 
 # TODO: Could potentially make this into set as well.
